created a few helper methods
